remove a bunch of compilation warnings
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@5753 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/pageitem.cpp
@@ -1272,7 +1272,7 @@ QString PageItem::ExpandToken(uint base)
 		QString out2;
 		//CB Section numbering
 		//out2 = out.arg(OwnPage+Doc->FirstPnum, -zae);
-		out2=out.arg(m_Doc->getSectionPageNumberForPageIndex(OwnPage), -zae);
+		out2=out.arg(m_Doc->getSectionPageNumberForPageIndex(OwnPage), -(int)zae);
 		//out2=out.arg(out2, -zae);
 		chstr = out2.mid(base-za2, 1);
 	}
